abstract reporting query
